renamed StatService to DataService to reflect its more general usage, and added interfaces.ts for all interface declarations
