+++ src/app/data.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { DataService } from './data.service';
+
+describe('DataService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [DataService]
+    });
+  });
+
+  it('should ...', inject([DataService], (service: DataService) => {
+    expect(service).toBeTruthy();
+  }));
+});
