--- src/app/idim.ts
@@ -1,5 +1,0 @@
-export interface Idim {
-    p1: [number, number],
-    p2: [number, number],
-    label: string
-}
