+++ src/app/interfaces.ts
@@ -1,0 +1,15 @@
+export interface Idim {
+    p1: [number, number],
+    p2: [number, number],
+    label: string
+}
+
+export interface Istat {
+    label: string;
+    value: number;
+  }
+
+export interface Ipoint {
+    x: number;
+    y: number;
+}
